+++ src/components/User/SignInForm.js
@@ -1,24 +1,13 @@
 import React, { useState } from 'react';
 import firebase from 'firebase/app';
 import SignUpForm from './SignUpForm';
-import {
-  GoogleLoginButton,
-  FacebookLoginButton,
-  TwitterLoginButton,
-} from 'react-social-login-buttons';
 import { Button, Form, FormGroup, Label, Input } from 'reactstrap';
 import styled from 'styled-components';
 import background from 'assets/img/splashscreen.png';
 import { ReactComponent as Logo } from 'assets/img/vota-cidade.svg';
 import Background from 'components/Background/Background';
 import { Tabs, TabButton } from 'components/Tabs/Tabs';
-
-const buttonStyle = {
-  fontSize: '12pt',
-  boxShadow: 'none',
-  paddingRight: 36,
-  width: '100%',
-};
+import SocialSignin from './SocialSignin';
 
 const Container = styled.div`
   height: 100vh;
@@ -63,7 +52,7 @@ const StyledSplashScreen = styled.div`
   justify-content: center;
 `;
 
-const SignInForm = (props) => {
+const SignInForm = ({ updateErrorMessage }) => {
   const [loading, setLoading] = useState(false);
   const [showSignUpForm, setShowSignUpForm] = useState(false);
   const [showForgotPasswordForm, setShowForgotPasswordForm] = useState(false);
@@ -80,63 +69,6 @@ const SignInForm = (props) => {
       setLoading(false);
     } catch (error) {
       console.log(error);
-    }
-  };
-
-  const signInWithFacebook = () => {
-    const provider = new firebase.auth.FacebookAuthProvider();
-    provider.addScope('email');
-
-    signInWithPopup(provider);
-  };
-
-  const signInWithGoogle = () => {
-    const googleProvider = new firebase.auth.GoogleAuthProvider();
-    googleProvider.addScope('profile');
-    googleProvider.addScope('email');
-    signInWithPopup(googleProvider);
-  };
-
-  const signInWithTwitter = () => {
-    const provider = new firebase.auth.TwitterAuthProvider();
-
-    signInWithPopup(provider);
-  };
-
-  const signInWithPopup = (provider) => {
-    firebase
-      .auth()
-      .signInWithPopup(provider)
-      .then(async (userCredential) => {
-        const { uid, displayName, email } = userCredential.user;
-      })
-      .catch(handleSocialLoginError);
-  };
-
-  const handleSocialLoginError = (error) => {
-    switch (error.code) {
-      case 'auth/account-exists-with-different-credential':
-        return props.updateErrorMessage(
-          'Esta conta já foi conectada usando outra rede social. Tente novamente usando outro método de login.',
-        );
-
-      case 'auth/popup-blocked':
-      case 'auth/popup-closed-by-user':
-        return props.updateErrorMessage(
-          'O popup de autorização foi bloqueado ou fechado, tente novamente.',
-        );
-
-      case 'auth/cancelled-popup-request':
-        return;
-
-      case 'auth/unauthorized-domain':
-      case 'auth/operation-not-supported-in-this-environment':
-      case 'auth/operation-not-allowed':
-      case 'auth/auth-domain-config-required':
-      default: {
-        console.error(error);
-        return props.updateErrorMessage('Ocorreu um erro inesperado');
-      }
     }
   };
 
@@ -203,26 +135,7 @@ const SignInForm = (props) => {
           </button>
         </StyledSpan>
         <Divider />
-        <div>
-          <FacebookLoginButton
-            onClick={signInWithFacebook}
-            text="Entrar com Facebook"
-            style={buttonStyle}
-            align="center"
-          />
-          <GoogleLoginButton
-            onClick={signInWithGoogle}
-            text="Entrar com Google"
-            style={buttonStyle}
-            align="center"
-          />
-          <TwitterLoginButton
-            onClick={signInWithTwitter}
-            text="Entrar com Twitter"
-            style={buttonStyle}
-            align="center"
-          />
-        </div>
+        <SocialSignin updateErrorMessage={updateErrorMessage} />
       </Box>
     </Container>
   );
